Add timeout control and unit tests for batch operations
This commit introduces a new "getRemainingTime" method in the OperationInterface that returns the number of seconds left before a timeout state has been reached. The timeout can be customized using the "max_execution_seconds" parameter in the "handleOperation" method in the Operator class. Also, a new DrupalBatchAPIOperationBaseTest.php file has been added to include corresponding unit tests for these changes.